+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased] - Next release 0.99.4
 
 ### Added
+- New query `getFlowNames` that returns a list of flow names grouped by flow type and can be filtered by flow state
 
 ### Changed